Single element Successor and Predecessor.

+++ Raven.Sparrow/Sparrow.Tests/BitVectors.cs
@@ -256,7 +256,7 @@ namespace Sparrow.Tests
 
             ulong[] data = new ulong[vectorSize];
             for (int i = 0; i < data.Length; i++)
-                data[i] = (ulong)rnd.Next() << 32 | (ulong)rnd.Next();
+                data[i] = (ulong)(rnd.Next() << 32 | rnd.Next());
 
             var v1 = BitVector.Of(data);
             var v2 = BitVector.OfLength(v1.Count);
